requirements.txt
requests==2.7.0
PyPDF2==1.26.0
pdfminer.six==20200726
beautifulsoup4==4.9.1